fix(emoji-blast-site): correct zero-coords for hero button kbd click
This adds the correction for zero coordinates for when the hero button
is clicked through the keyboard press (Enter key). There is no proper
way to detect when the "click" MouseEvent is triggered by the keyboard,
despite handling the zero coords, which also not always zero (Safari
should set them to the center of the clicked element (for keyboard
"clicks")).